Proper refactoring of classes, packaging, javadoc, code style
